feat(masthead-contact): update CMApp interface
